requirements.txt
aniso8601==9.0.1
attrs==24.2.0
autopep8==2.3.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
Flask==3.1.0
flask-restx==1.3.0
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.1
pycodestyle==2.12.1
PyMatting==1.1.13
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.13.1
sympy==1.13.3
tifffile==2024.8.30
tomli==2.2.1
tqdm==4.67.1
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
